requirements.txt
begins==0.9
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
idna==2.7
imageio==2.4.1
moviepy==0.2.3.5
numpy==1.15.3
Pillow==5.3.0
pytube==9.3.3
requests==2.20.0
tqdm==4.28.1
urllib3==1.24
